# examples/todo-api/openapi.yaml
openapi: "3.0.3"
info:
  title: Todo API
  version: "1.0.0"
  description: A simple todo list API for testing mcp-c
servers:
  - url: http://localhost:4000
tags:
  - name: todos
    description: Manage todo items
  - name: tags
    description: Manage tags for todos
paths:
  /todos:
    get:
      operationId: listTodos
      summary: List all todos
      tags: [todos]
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by status
          schema:
            type: string
            enum: [pending, done, all]
        - name: limit
          in: query
          required: false
          description: Max items to return
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      summary: Create a new todo
      tags: [todos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  description: Todo title
                description:
                  type: string
                  description: Detailed description
                priority:
                  type: string
                  description: Priority level
                  enum: [low, medium, high]
                  default: medium
                tags:
                  type: string
                  description: Comma-separated tags
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    get:
      operationId: getTodo
      summary: Get a todo by ID
      tags: [todos]
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      responses:
        "200":
          description: Todo details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    put:
      operationId: updateTodo
      summary: Update a todo
      tags: [todos]
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Todo title
                description:
                  type: string
                  description: Detailed description
                status:
                  type: string
                  description: Todo status
                  enum: [pending, done]
                priority:
                  type: string
                  description: Priority level
                  enum: [low, medium, high]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Todo updated
    delete:
      operationId: deleteTodo
      summary: Delete a todo
      tags: [todos]
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Todo deleted
  /tags:
    get:
      operationId: listTags
      summary: List all available tags
      tags: [tags]
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, done]
        priority:
          type: string
          enum: [low, medium, high]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer